+++ model/sector.go
@@ -1,7 +1,6 @@
 package model;
 
 type Sector struct {
-
 	Name string `json: "name"`
 
 }